requirements.txt
fastapi
uvicorn
pydantic
requests
langchain
langchain-community
chromadb
sentence-transformers
groq
python-dotenv
streamlit